Move stateful offset management responsibility from fragment-helper to coordinator-wrapper
